moved CanSort & CanFilter from TableViewBoundColumn to TableViewColumn
